+++ src/NavigatorLazy.tsx
@@ -42,7 +42,7 @@ import IFrame from './components/IFrame';
 import WithNavigation from './components/WithNavigation';
 import Registration from './pages/Registration';
 
-import RequestMatchStudent from './pages/student/matching_new/RequestMatch';
+import RequestMatchStudent from './pages/student/matching/RequestMatch';
 import RequestMatch from './pages/pupil/matching/RequestMatch';
 import Matching from './pages/pupil/Matching';
 import NotficationControlPanel from './pages/notification/NotficationControlPanel';
